Irwin validation

+++ TerraformMinds/Controllers/AdministratorController.cs
@@ -76,7 +76,7 @@ namespace TerraformMinds.Controllers
                 ViewBag.GradeLevels = gradeLevelList;
 
                 ViewBag.GradeLevel = gradeLevel;
-                ViewBag.CurrentCapacity = currentCapacity;
+                //ViewBag.CurrentCapacity = currentCapacity;
                 ViewBag.MaxCapacity = maxCapacity; 
             }
             return View();
@@ -106,10 +106,16 @@ namespace TerraformMinds.Controllers
         // more details, see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost, ActionName("CourseCreate")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> CourseCreate(string instructor, string courseName, string subject, string courseDescription, string gradeLevel, DateTime startDate, DateTime endDate, int currentCapacity, int maxCapacity, [Bind("ID,UserID,CourseName,Subject,CourseDescription,GradeLevel,StartDate,EndDate,CurrentCapacity,MaxCapacity")] Course course)
+        public async Task<IActionResult> CourseCreate(string instructor, string courseName, string subject, string courseDescription, string gradeLevel, DateTime startDate, DateTime endDate, int currentCapacity, int maxCapacity, [Bind("ID,UserID,CourseName,Subject,CourseDescription,GradeLevel,StartDate,EndDate,MaxCapacity")] Course course)
         {
             using (LearningManagementContext context = new LearningManagementContext())
             {
+
+                instructor = instructor != null ? instructor.Trim() : null;
+                courseName = courseName != null ? courseName.Trim() : null;
+                subject = subject != null ? subject.Trim() : null;
+                courseDescription = courseDescription != null ? courseDescription.Trim() : null;
+
                 if (ModelState.IsValid)
                 {
                     try
@@ -127,9 +133,17 @@ namespace TerraformMinds.Controllers
 
                             else
                             {
-                                if(context.Courses.Any(x => x.CourseName == course.CourseName))
+                                if (course.StartDate < DateTime.Today)
                                 {
-                                    exception.ValidationExceptions.Add(new Exception("Invalid: That course already exists."));
+                                    exception.ValidationExceptions.Add(new Exception("Invalid: Course Start Date cannot be set in the past."));
+                                }
+
+                                else
+                                {
+                                    if (context.Courses.Any(x => x.CourseName == course.CourseName))
+                                    {
+                                        exception.ValidationExceptions.Add(new Exception("Invalid: That course already exists."));
+                                    }
                                 }
                             }
                         }
@@ -173,7 +187,7 @@ namespace TerraformMinds.Controllers
                     ViewBag.GradeLevels = gradeLevelList;
 
                     ViewBag.GradeLevel = gradeLevel;
-                    ViewBag.CurrentCapacity = currentCapacity;
+                    //ViewBag.CurrentCapacity = currentCapacity;
                     ViewBag.MaxCapacity = maxCapacity;
                 }
             }
@@ -288,6 +302,10 @@ namespace TerraformMinds.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> CourseEdit(int id, [Bind("ID,UserID,CourseName,Subject,CourseDescription,GradeLevel,StartDate,EndDate,CurrentCapacity,MaxCapacity")] Course course)
         {
+            course.CourseName = course.CourseName != null ? course.CourseName.Trim() : null;
+            course.Subject = course.Subject != null ? course.Subject.Trim() : null;
+            course.CourseDescription = course.CourseDescription != null ? course.CourseDescription.Trim() : null;
+
             using (LearningManagementContext context = new LearningManagementContext())
             {
                 if (id != course.ID)
@@ -358,7 +376,6 @@ namespace TerraformMinds.Controllers
                     ViewBag.GradeLevel = course.GradeLevel;
                     ViewBag.CurrentCapacity = course.CurrentCapacity;
                     ViewBag.MaxCapacity = course.MaxCapacity;
-                    //return RedirectToAction(nameof(CourseList));
                 }   
             }
             return View(course);
@@ -394,11 +411,37 @@ namespace TerraformMinds.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             using (LearningManagementContext context = new LearningManagementContext())
-            {
+            {                
                 var course = await context.Courses.FindAsync(id);
-                context.Courses.Remove(course);
-                await context.SaveChangesAsync();
-                return RedirectToAction(nameof(CourseList));
+
+                try
+                {
+                    if (course.StartDate < DateTime.Today)
+                    {
+                        exception.ValidationExceptions.Add(new Exception("Error: Cannot delete a course that is in progress."));
+                    }
+
+                    else if (course.EndDate > DateTime.Today && course.StartDate < DateTime.Today)
+                    {
+                        exception.ValidationExceptions.Add(new Exception("Error: Cannot delete a course that is in progress."));
+                    }
+
+                    if (exception.ValidationExceptions.Count > 0)
+                    {
+                        throw exception;
+                    }
+                    context.Courses.Remove(course);
+                    await context.SaveChangesAsync();
+                    return RedirectToAction(nameof(CourseList));
+                }
+
+                catch (ValidationException e)
+                {
+                    ViewBag.Message = "There exist problem(s) with your submission, see below.";
+                    ViewBag.Exception = e;
+                    ViewBag.Error = true;
+                }
+                return View(course);
             }
         }
